Refactor NotesOverview
