+++ apps/back-office/src/app/melding/[meldingId]/notities/NotesOverview.test.tsx
@@ -2,13 +2,7 @@ import { render, screen } from '@testing-library/react'
 
 import type { NoteRetrieveOutput } from '@meldingen/api-client'
 
-import { formatDateTime, NotesOverview } from './NotesOverview'
-
-describe('formatDateTime', () => {
-  it('formats a date string correctly', () => {
-    expect(formatDateTime('2024-03-05T08:09:00Z')).toBe('05-03-2024 09:09')
-  })
-})
+import { NotesOverview } from './NotesOverview'
 
 const defaultProps = {
   currentUserId: 1,
@@ -28,96 +22,26 @@ describe('NotesOverview', () => {
     expect(title).toBeInTheDocument()
   })
 
-  it('renders notes when provided', () => {
+  it('renders a list of notes', () => {
     const notes = [
-      { created_at: '2024-03-05T14:07:00Z', id: 1, text: 'This is a test note.', user: { email: 'test@example.com' } },
       {
-        created_at: '2024-03-05T15:07:00Z',
+        created_at: '2024-03-05T14:07:00Z',
+        id: 1,
+        text: 'This is a test note.',
+        user: { email: 'test@example.com' },
+      },
+      {
+        created_at: '2024-04-05T14:07:00Z',
         id: 2,
         text: 'This is another test note.',
-        user: { email: 'another@example.com' },
+        user: { email: 'test@example.com' },
       },
     ] as NoteRetrieveOutput[]
 
     render(<NotesOverview {...defaultProps} notes={notes} />)
 
     expect(screen.getByText('This is a test note.')).toBeInTheDocument()
-    expect(screen.getByText('test@example.com')).toBeInTheDocument()
-    expect(screen.getByText('05-03-2024 15:07')).toBeInTheDocument()
-
     expect(screen.getByText('This is another test note.')).toBeInTheDocument()
-    expect(screen.getByText('another@example.com')).toBeInTheDocument()
-    expect(screen.getByText('05-03-2024 16:07')).toBeInTheDocument()
-  })
-
-  it('does not show an edit link for notes that do not belong to the current user', () => {
-    const notes = [
-      {
-        created_at: '2024-03-05T14:07:00Z',
-        id: 1,
-        text: 'This is a test note.',
-        user: { email: 'test@example.com', id: 2 },
-      },
-    ] as NoteRetrieveOutput[]
-
-    render(<NotesOverview {...defaultProps} notes={notes} />)
-
-    const editLink = screen.queryByRole('link', { name: 'edit-link' })
-
-    expect(editLink).not.toBeInTheDocument()
-  })
-
-  it('shows an edit link for notes that belong to the current user', () => {
-    const notes = [
-      {
-        created_at: '2024-03-05T14:07:00Z',
-        id: 1,
-        text: 'This is a test note.',
-        user: { email: 'test@example.com', id: 1 },
-      },
-    ] as NoteRetrieveOutput[]
-
-    render(<NotesOverview {...defaultProps} notes={notes} />)
-
-    const editLink = screen.getByRole('link', { name: 'edit-link' })
-
-    expect(editLink).toBeInTheDocument()
-  })
-
-  it('shows a deleted note message when the note text is empty', () => {
-    const notes = [
-      { created_at: '2024-03-05T14:07:00Z', id: 1, text: '', user: { email: 'test@example.com' } },
-    ] as NoteRetrieveOutput[]
-
-    render(<NotesOverview {...defaultProps} notes={notes} />)
-
-    expect(screen.getByText('deleted-note')).toBeInTheDocument()
-  })
-
-  it('lets a user know when a note was edited', () => {
-    const notes = [
-      {
-        created_at: '2024-03-05T14:07:00Z',
-        id: 1,
-        text: 'This is an edited test note.',
-        updated_at: '2024-03-05T15:07:00Z',
-        user: { email: 'test@example.com', id: 2 },
-      },
-      {
-        created_at: '2024-03-05T16:07:00Z',
-        id: 2,
-        text: 'This note was not edited.',
-        updated_at: '2024-03-05T16:07:00Z',
-        user: { email: 'another@example.com', id: 3 },
-      },
-    ] as NoteRetrieveOutput[]
-
-    render(<NotesOverview {...defaultProps} notes={notes} />)
-
-    const visualOnlyEditedText = screen.getByText('edited', { exact: true, selector: 'span' })
-    const visuallyHiddenEditedText = screen.getByText('visually-hidden-texts.edited')
-
-    expect(visualOnlyEditedText).toBeInTheDocument()
-    expect(visuallyHiddenEditedText).toBeInTheDocument()
+    expect(screen.getAllByText('test@example.com')).toHaveLength(2)
   })
 })
